backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.5",
    "anthropic>=0.69.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.14.2",
    "boto3>=1.40.42",
    "celery[redis]>=5.5.3",
    "cryptography>=46.0.2",
    "fastapi[standard]>=0.118.0",
    "httpx>=0.28.1",
    "langgraph>=0.6.8",
    "langgraph-checkpoint>=2.1.1",
    "openai>=2.0.0",
    "playwright>=1.55.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.43",
    "supabase>=2.20.0",
    "svix>=1.76.1",
    "upstash-redis>=1.4.0",
    "uvicorn[standard]>=0.37.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.2",
]
